fixed que create option's positon issue and other spacing issues
