+++ src/dashboard/components/molecules/createquestion/QuestionCreateForm.jsx
@@ -312,7 +312,7 @@ export default function QuestionCreateForm() {
             )}
           </div>
 
-          <div className="flex flex-col md:flex-row items-center justify-between gap-4 space-y-1 pb-10">
+          <div className="flex flex-row items-center justify-between gap-4 space-y-1 pb-4">
             {/* is_paid */}
             <div>
               <Checkbox
