+++ src/features/questions/questionsCategoryApi.js
@@ -6,6 +6,10 @@ export const questionsCategoryApi = apiSlice.injectEndpoints({
             query: (type) => `/questions/${type}`,
             providesTags: ["Questions"],
         }),
+        getSingleCategory: builder.query({
+            query: ({type, id}) => `/questions/${type}/${id}`,
+            providesTags: ["Questions"],
+        }),
         createQuestionsCategory: builder.mutation({
             query: ({ type, data }) => ({
                 url: `/questions/${type}`,
@@ -35,6 +39,7 @@ export const questionsCategoryApi = apiSlice.injectEndpoints({
 export const { 
     useCreateQuestionsCategoryMutation, 
     useGetQuestionsCategoryQuery, 
+    useGetSingleCategoryQuery,
     useDeleteQuestionsCategoryMutation,
     useEditQuestionsCategoryMutation 
 } = questionsCategoryApi;
